[RFC] Support paging iterator for bulk fetch requests.
Previously paging was handled internally and all pages
were collected on each request.

This change exposes an iterator of the pages of responses
that followed the link headers returned on each request.

This does not directly change any behavior of the Mylyn
GitHub connector since issues and gist queries still return
the full response.

Change-Id: I8c12e20e4115e1968949319b3e9214e06db30b8b
Signed-off-by: Kevin Sawicki <kevin@github.com>