Update patch set 4
Patch Set 4:

no, it's just a convenient way to bring all the needed dependencies into your workspace and this is what the contributor guide recommends [1] hence I tried this. In EGit so far we don't have a shared target definition so the set of dependencies is resolved against the Eclipse instance used for development + the projects in the workspace. A target definition is a more precise way to define what is the target environment so maybe we should also define that for EGit. I'll look into that.

[1] http://wiki.eclipse.org/EGit/GitHub/ContributorGuide

Patch-set: 4