Remove use of Dequeue from DateFormatter
I came across an issue where Dequeue was not available in Android distributions
earlier than Gingerbread, which caused crashes of my app when Dequeue was not
found. Because it doesn't appear that DateFormatter actually requires the
functionality of Dequeue, I've replaced its usage with LinkedList.

I've tested this in my application and the change does not appear to have any
adverse effects.

Change-Id: I4baf4c71945994d53966673c95819b3221bb6e6e
Signed-off-by: Eddie Ringle <eddie@eringle.net>