Adds a method to edit a repository
The editRepository method is missing from the RepositoryService.
This patch adds that method. Since the createRepository live test
is ignored I was uncertain how to best add a test that performs
modifications. Ideally, this should be done after a repository has
been added and before it is removed again. There seem to be no way
to remove a repository through the API though.

Change-Id: Ib69f15094d568d015ab36706f553d337a73c984c