Update patch set 4
Patch Set 4:

Thanks for the contribution, I made a few minor edits to the commit message format and added a Signed-off-by line for you.

I also added a second unit test that verifies a POST to the correct URL.

Please let me know if you are okay with these changes.

Patch-set: 4