[RFC] Support configuration of alternate client implementations
The default client uses Apache HttpComponents (HC) which is not
entirely suitable for Android deployments since it requires
distributing a shaded JAR with custom logging configuration.

A new base class HttpClient has been introduced that can
be implemented to use different HTTP libraries as well as the
built-in HttpURLConnection.

This commit removes some use of HC as well as introducing base
classes that allow non-HC implementations to exist.  Services
still default to using a GitHubClient for backwards compatibility
but the goal is to be able to use the entire GitHub Java API library
without loading a single class from HC.

Change-Id: I82e442139d3adec3d2331f279bde69320b13638d
Signed-off-by: Kevin Sawicki <kevin@github.com>