Use custom deserializer for Event class
Using newer streaming versions of Gson causes deserialization
problems if the payload is before the type field in the returned
JSON using the current custom Event formatting.

This changes the behavior to use an internal custom Gson instance
just to handle parsing events that allows a chance to check the
type and create the custom payload before the Event object is
returned to any service callers.

This approach simplifies the code but does require an additional
Gson instance to be kept around to handle the default serialization
of the top-level Event fields.

Change-Id: Ic10a755bbf798ec2d1d4a4c53ccd9154e0cb91e2