Show info message that anonymous cannot submit tasks
When creating or refreshing the editor, it is checked if the github
repository is configured with user credentials. In case, it is anonymous
an info message is shown. When clicking on the info message, the
repository configuration dialog is opened and the user can enter the
credentials.

If the repository is anonymous and the user insists to submit the issue,
then the info message turns into error message.

This solution is inspired by the Bugzilla task editor.

Bug: 390757
Change-Id: I9f498acdec46a15156939e71bcc9bd40766cd65f