Bug 398857 - Allow requiring only one of selected labels for issue queries [DRAFT]
Added option to require an issue have one or more of the chosen labels
for a task query, rather than requiring all labels.  Added name and
values for a query attribute to IssueService.  Added appropriate radio
buttons to IssueRepositoryQueryPage, and messages to
Messages/messages.properties.  Changed IssueConnector to handle the new
option: if all labels are required, query is run as before, otherwise it
is run for each individual label.

Change-Id: Id7887be84339d9b91f53713f25e4744daf1881ea