Fix the auth failure issue when working with GHE
The UserPlan object on github enterprise returns very large numeric
values for certain fields that do not fit in an int. This causes a
deserialization error when GSON is parsing the auth response.

The serialVersionUID has also been updated in this patch as the types of
member variables has undergone a change

Bug: 399003
Change-Id: I8082f50d52902d3ceda5f82a412454ceb44b2872