Update patch set 1
Patch Set 1:

> What about the `rename` attribute for "rename" events? (which gives
 > the old and new titles of the issue)

Yes, the `rename` attribute is the only one missing, as it needs a new object and I wanted my first contribution to be small (just to get used to the process). If my patch gets accepted I'll keep on working on this kind of stuff.

 > (1 comment)

Thanks for the review, I'll do the change locally and wait until a committer confirms if it's needed or not.

Patch-set: 1
Reviewer: Gerrit User 15912 <15912@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0