Remove deprecated access to mylyn.tasks
Now using new (redirected) values of several deprecated constants.
Minimum version of mylyn.tasks.core has been increased to the version
where the new constants are declared. That minimum version is years
behind the current version, so this does not really seem unfair to users
of mylyn, as it does not enforce upgrading to a very recent version.

Checked with egit-4.5 target platform.

Change-Id: If49165fcaf4956079935a091887da6e5ad412316
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>