Fix GitHubClient to always use responseContentType
The value of the GitHubRequest responseContentType field is supposed to be
used as the HTTP Accept header for that particular GitHubRequest. This field
exists because various endpoints in the GitHub API support multiple response
media types. See: https://developer.github.com/v3/media/.

This field is respected by GitHubClient's get(GitHubRequest) method, but this
is *not* the case in getStream(GitHubRequest). The handling of the
responseContentType field between these two very similar methods should
definitely be consistent.

Bug: 528982
Change-Id: Ie7964bf2337b4547ea602b0bfa3f2c20afc8df83
Signed-off-by: Christian Barcenas <christian@cbarcenas.com>