Update patch set 1
Patch Set 1: Code-Review+1

This is OK, but doing the multiplicity distinctions in code is bad for i18n. There are languages that have highly irregular plural forms, where the words for 0, 1, 2, 3 or also 10, 11, 12 (and 20, 21, 22 etc.) are different. Thus it's in general better to make the whole plural form localizable. For instance a single message

  RepositorySearchWizardPage_counters=({0,choice,0#No forks|1#One fork|1<{0} forks}, {1,choice,0#No watchers|1#One watcher|1<{1} watchers})

used via

  MessageFormat.format(Messages.RepositorySearchWizardPage_counters, repo.getForks(), repo.getWatchers());

enables a translator to use grammatically correct plural forms in all cases.

Since you're touching that code anyway, why not refactor it in that way?

Patch-set: 1
Label: Code-Review=+1