Update patch set 1
Patch Set 1:

> Michael, have you seen the comment above? Somehow I can't seem to
 > manage to enable PDE API analysis properly and make it emit errors
 > when those @since annotations are not there. Any idea?

As far as I can see, a workspace can have only one API baseline, which is egit-5.1.1 in the preferences. That does not contain any of the github plugins. That may be the reason. At least you should see a warning for the missing baseline for each github plugin, if you start enabling the API analysis on those projects via project preferences.

Or is this about how get the API analysis properties at all? They are in the "Plugin tools" context menu of the projects.

Patch-set: 1