Update patch set 1
Patch Set 1:

u should also think about storing the original SVG somewhere.
For platform projects we have a separate repo for this.
Maybe you also wanna do this for egit. Or just put them into a separate folder in the plugin (that is not copied into the plugins's jar)

Patch-set: 1
Label: Code-Review=0