Update patch set 3
Patch Set 3:

> (1 comment)

So just FYI I have posted a question to github api forums to see if I can get an authoritative answer instead of just me trying a few things and drawing a conclusion: https://github.community/t5/GitHub-API-Development-and/Does-Github-API-support-X-HTTP-Method-Override/m-p/57361

I'll hold off until we get some type of answer.

Another long shot option would be this: Could we create an optional org.eclipse.egit.github.core.patch OSGi fragment module that could have BREE set to JDK11 and in that fragment we could put a delegate class that could handle the patch method?   And then in GitHubClient, it could do a Class.forName("") and if the fragment module is online (because BREE is available) then forName() will return with a valid delegate class that we could call?   Is that way over-engineered? :)

Patch-set: 3