Rename category and add repository reference to egit update site
Rename the category from generic "Collaboration" which is used in the
p2 repository of the simultaneous release to a more specific name
EGit.Github and give a more precise description for this category.

Adding the reference to the egit update site should help users having
egit and egit github integration installed and trying to update to a
newer release using the github p2 repository and have not checked
"Contact all update sites".

Change-Id: If6298390b2dae07f6368a4884a0261e65676ade0
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>