[releng] Force --release and use BREE to run tests
Make the EGit build work if code is compiled with a Java > 1.8.
This prepares for moving the BREE to Java 11: when we do that, we
will need to be able to build nightly with and against Java 11, but
the stable branch against Java 1.8.

This requires a ~/.m2/toolchains.xml file defining at least a
JavaSE-1.8 toolchain. The CI build pods do have such toolchain
definitions.

Change-Id: I82e9cb4c2518bc13906d44cc1ead2724f7130ffe
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>