Switch BREE to Java 11
Update BREE, use --release 11, set target based on Eclipse 2020-09,
which was the first Eclipse version that required Java 11. This is
the same as the baseline of EGit. Update JDT settings. Remove pack200.

Bump plugin-versions to the newest:

* jarsigner 1.3.2
* maven-pmd 3.15.0
* spotbugs 4.4.1
* tycho 2.5.0

Suppress *EXPOSE_REP* warnings from spotbugs; see [1].

Remove two profiles that only set a variable that was unused anyway.

[1] https://github.com/spotbugs/spotbugs/issues/1601

Bug: 576238
Change-Id: If5ab0eca5c9e4d92355ec314b18ea95fd4d6cdf6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>